requirements.txt
Put all requirements in here once I know all the proper libraries
python
blockfrost-python